refactor: rename cabin module
Proposing name change from cabin to cabinUt in all relevant names to make it clearer that these refer to cabins from UT.
