+++ src/bot/bot.service.ts
@@ -3,7 +3,7 @@ import { DiscordService } from 'src/discord/discord.service';
 import * as cron from 'node-cron';
 import { convert } from 'html-to-text';
 import { EmbedBuilder } from 'discord.js';
-import { CabinService } from 'src/cabin/cabin.service';
+import { CabinUtService } from 'src/cabinUt/cabinUt.service';
 import { ConfigService } from '@nestjs/config';
 import { ActivityDatabaseService } from 'src/activityDatabase/activityDatabase.service';
 import { SupabaseRealtimePayload } from '@supabase/supabase-js';
@@ -15,7 +15,7 @@ export class BotService implements OnModuleInit {
   constructor(
     private readonly configService: ConfigService,
     private readonly discordService: DiscordService,
-    private readonly cabinService: CabinService,
+    private readonly cabinUtService: CabinUtService,
     private readonly activityDatabaseService: ActivityDatabaseService,
   ) {
     this.newActivityChannelId = this.configService.getOrThrow(
@@ -70,7 +70,7 @@ export class BotService implements OnModuleInit {
     const channelId = this.configService.getOrThrow(
       'DISCORD_DAILY_CABIN_CHANNEL_ID',
     );
-    const cabin = await this.cabinService.getRandomCabin();
+    const cabin = await this.cabinUtService.getRandomCabin();
     const text = convert(cabin.description);
 
     const imageUrl = `https://res.cloudinary.com/ntb/image/upload/w_1280,q_80/v1/${cabin.media[0].uri}`;
