+++ src/discord/discord.module.ts
@@ -1,11 +1,11 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { DiscordService } from './discord.service';
 import { ConfigModule } from '@nestjs/config';
-import { CabinModule } from 'src/cabin/cabin.module';
+import { CabinUtModule } from 'src/cabinUt/cabinUt.module';
 import { SlashCommandModule } from 'src/slashCommand/slashCommand.module';
 
 @Module({
-  imports: [ConfigModule, CabinModule, forwardRef(() => SlashCommandModule)],
+  imports: [ConfigModule, CabinUtModule, forwardRef(() => SlashCommandModule)],
   providers: [DiscordService],
   exports: [DiscordService],
 })
